generalize sync status API
